# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lighthouse",
    "description": "Knowledge base for AI agents. Public retrieval over HTTP (MCP wrapper on top), proposal pipeline behind a shared API key.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/search": {
      "get": {
        "tags": [
          "retrieval"
        ],
        "summary": "Search",
        "operationId": "search_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Natural-language query",
              "title": "Q"
            },
            "description": "Natural-language query"
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fetch_entity/{node_id}": {
      "get": {
        "tags": [
          "retrieval"
        ],
        "summary": "Fetch Entity",
        "operationId": "fetch_entity_v1_fetch_entity__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fetch_source/{episode_id}": {
      "get": {
        "tags": [
          "retrieval"
        ],
        "summary": "Fetch Source",
        "operationId": "fetch_source_v1_fetch_source__episode_id__get",
        "parameters": [
          {
            "name": "episode_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/corpus/sources": {
      "get": {
        "tags": [
          "v1",
          "corpus"
        ],
        "summary": "Sources",
        "description": "Per-source roll-up. `order=chunks` ranks by chunk count\n(largest sources first); `order=recent` by most recent ingest.",
        "operationId": "sources_v1_corpus_sources_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(chunks|recent)$",
              "default": "chunks",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CorpusSource"
                  },
                  "title": "Response Sources V1 Corpus Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CorpusSource": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "recipes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recipes"
          },
          "last_ingest_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ingest At"
          }
        },
        "type": "object",
        "required": [
          "source",
          "chunk_count",
          "recipes",
          "last_ingest_at"
        ],
        "title": "CorpusSource"
      },
...
      "EntityResponse": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "name",
          "summary"
        ],
        "title": "EntityResponse",
        "description": "One entity returned by ``/fetch_entity``."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchHit": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Node Id"
          },
          "target_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Node Id"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "episode_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Episode Ids"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "summary"
        ],
        "title": "SearchHit",
        "description": "One fact returned by ``/search``.\n\nA fact is a graph edge \u2014 a one-line statement relating two entities\n(``source_node_id`` and ``target_node_id``) extracted from a source\nchunk (``episode_ids``). Clients that want more context either\ndrill into an entity via ``/fetch_entity`` or \u2014 usually preferred \u2014\npull the original paragraph via ``/fetch_source``."
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "type": "array",
            "title": "Hits"
          }
        },
        "type": "object",
        "required": [
          "query",
          "hits"
        ],
        "title": "SearchResponse"
      },
      "SourceResponse": {
        "properties": {
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "valid_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid At"
          }
        },
        "type": "object",
        "required": [
          "episode_id",
          "name",
          "source",
          "content"
        ],
        "title": "SourceResponse",
        "description": "One source chunk returned by ``/fetch_source``."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Lighthouse-Key"
      }
    }
  }
}